Adding priority system for organizations.
This commit adds a new optional configuration variable, PRIORITY, for
a team. The accepted values at first are 'primary' or 'secondary', but
in reality it only is used for an alphabetical sort additional when
sorting organizations, plus some view work to pull out the 'secondary'
entries with their own heading.

The idea here is that if your company had many organizations, you
might want to have the main orgs listed first, and then a list of
orgs which have less emphasis afterwards, for more specialized
projects.

Once we hit about 6 orgs this felt necessary to move beyond just the
primary alphabetical sort.