Supporting organizations that do not need repo approvers.
If there is no repo approvers team assigned to an organization, we
will make the assumption that this org lets users create repos as
a general member privilege.

In the absense of a defined approver team ID, we will close off the
repo approval UI for users, and instead use the error system to show
a basic message to the user, including a link off to GitHub to
create a repo using the URL on GitHub for an org's new repo page.