Improving the error page display.
This updates the error system to show a little less information,
formatted differently, when we know that this is an expected
error message (i.e. a validation error, or a message intended for
a user, but not actually an error).

When the `skipLog` param is set on an error object, the display
on the error template will change: it will be a little less
scary, and the Report Bug button will instead become a button to
communicate with the admins, but implying that this is not an
error.

Also, adding a "Go Back" button to the page, using the JavaScript
`history.back()` function, when we know that the error is user-
indented.