Adding lock support for organizations.
This is not a very great feature for general use, but it can help
while testing and onboarding new organizations.

A "locked" organization will not allow anyone to join the org through
the portal.

In time, I'd like to add an experience whereby there is an approval
process for organizations that not everyone that is an employee should
have access to.

This comes down to departmental orgs: there's risk and no good reason
in letting everyone join.

Ahead of that workflow being available, locking is a way to let people
use parts of the portal if they are already members, but to not allow
new joins through this oss portal.