Passport Azure AD: handling special scenario when opening hyperlinks We identified an issue where opening a hyperlink through certain applications on Windows, such as Word, may actually open a browser through the operating system in a way that there is no way to have an initial valid session cookie. As a result, a generic 'Unauthorized' message was being seen in this case.
This fix monkey patches the Passport Azure AD library to intercept
a specific failure message and instead to use that to redirect the
user instead of showing the failure. We can then show a pretty message,
or, if the browser has then established a proper cookie (the common
case), we just redirect to the homepage of the app to let the user at
least still try to use the app.