bugfix: membership publicize
The publicize/conceal operation on a GitHub organization requires an
active GitHub user session including an enhanced write-scoped token.

In the original OAuth world, this required a new OAuth dance; however,
with a GitHub App, we just need the token available. Adding the existing
middleware to require an active GitHub session for the membership page
helps achieve this.