New repo emails: notify the manager
If a manager is known for an employee, we will also notify
the manager via a carbon copy on e-mails related to new repos,
locked forks, or newly created repos that still need classification.

As a quick change, this does not add a feature flag to enable or disable
this; however, if a manager is not known via the corporate contacts or
cached employee information system, there will be no error.