New feature: self-service team maintainer upgrade
This introduces a new opt-in (feature flag) capability to reduce the
amount of manual organization operations that are required for
common scenarios.

For a relatively small GitHub team size (configurable, default to
around 15 people), this allows people to self-service "upgrade"
themselves to be a team maintainer if the team does not have
many maintainers.

The # of maintainers that are supported by default for this feature
to light up is zero (0) or 1... since we recommend that every
team have at least 2 maintainers.