Native Entities refactoring (Postgres)
Refactoring of the Postgres and EMP (entity metadata provider) interface
to make it more performant, moving beyond the original key-value store
approach for entities and allowing for "native" columns that can have
solid indexes on the columns.

The goal of this integration is to be able to create indexes on the
date fields for open source contributions. Previously, storing this
data inside of Postgres JSON columns made casting difficult or slow
to compute when querying for these values.

Also begins to ship queries away from a central cross-EMP query enum
into compartmenatlized entities.