June updates
During the month of May and June, we forked this application internally
as part of a specific set of temporary changes. To converge once again
with the open source version and make it easier to collaborate with others,
this brings the latest changes up.

There is a need to refactor and feature flag the service account piece into
the newer experience as new debt.