Support "main" default branch for new repos
These changes allow new repos to take on the `main` default branch name
in this interim period before the defaults change on GitHub.

For any given GitHub org that is managed, *if* the org has had the Updates GitHub app
installed and configured as an installation, the rename operations can operate on
repo contents and succeed. Otherwise, the branch default will not rename.

Also, API-created repos (by the docs.msft team today) are not adjusted by configuration.

It may be worth adding a dynamic setting for orgs to opt-in, but for now, this is a validation
pass, as only Updates-installed orgs like the Microsoft org will take on the capability.