This will eventually enable an "easy button" experience to rename a branch to `main`, or another name, including rewriting protected branch restriction rules, creating the branch, updating any open pull requests targeting the old branch name, and removing the old branch, with complete logging.
Based on work done by GitHub and others, adapted to our environment,
and improved error logging.

Adds support for GraphQL queries to the GitHub interface system.

This feature requires that an org has had an installation of the "Updates"
app which has access to read/write branches, pull requests, file contents,
as well as administer a repo.