Organization sudo security group support
Adds support for organization sudo privileges to flow from
security groups instead of from GitHub Teams as the source of
truth.

Sudo remains on by default and configured for teams.

Also provides for company-specific overrides if you have a
different system for authorization decisions that is not part
of the default kit.