Portal sudo SG support
Supports using security groups for portal sudo permissions
instead of the primary GitHub org's sudo configuration, if
present.

Breaking feature change:
This feature is not enabled by default and must use a feature flag
to opt-in.

The feature flag is "FEATURE_FLAG_ALLOW_PORTAL_SUDO".

The standard behavior, when the flag is enabled, is still to use
the first org. Alternatively, configuration and a SG ID can be provided.