New repo lockdown: identity accidental co-2-co forks
We often find that teams fork a repo from another of our "official" GitHub orgs
by accident, often as they are still new to the fact that they should probably be
contributing simple pull requests from a fork in their individual GitHub account,
and not a company-official org.

We see this when people fork a repo like MicrosoftDocs/azure-docs into the official
Microsoft org, yielding Microsoft/azure-docs.

This totally messes with the search engines when people start finding the fork instead
of the upstream parent, etc.

This helps identify the case, provide a custom message when configured, etc.