GitHub Apps: do not enforce all repo installs
To support specific scenarios such as installing one of the sub-apps in just
a repo or two, we should no longer throw on a GitHub app that is not installed
in all repos in the target organization.

This is an interim improvement for the scenario but should be specialized in
the future to still catch situations where an official app was only partially
installed.