Organization: get Copilot seat assignments
Retrieves the list of Copilot basics when the organization
has Copilot for Business attachments.

Do note that this API only works to page, slowly, at scale to
a point - after 400 pages I am not able to get results due to
a hard cap of sorts here, so need to rely on other methods to
get the info.